move sample table column defs into their own files
